requirements.txt
# Core
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
httpx>=0.24.0
aiohttp>=3.9.0
feedparser>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
aiosqlite>=0.19.0

# Math/Stats
scipy>=1.11.0
numpy>=1.24.0

# Monitoring
datasette>=0.64.0
datasette-auth-passwords>=0.4.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.1.0
pyyaml>=6.0.0

# Polymarket
py-clob-client>=0.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
